requirements.txt
pillow>=10.0.0  # For image processing
tkinter  # Usually comes with Python installation
json # For reading and writing JSON output files
time # For better timing 
altgraph                  
imageio                   
nuitka                    
numpy                     
ordered-set               
packaging                 
pefile                    
pillow                    
pip                       
pygame                    
pyinstaller               
pyinstaller-hooks-contrib 
pywin32-ctypes            
setuptools                
zstandard                 